requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
python-dotenv==1.0.0
pytest==8.3.5
pytest-asyncio==0.26.0
httpx==0.25.2
replicate==1.0.6 